+++ apps/api/gql/schema/__mocks__/graphql.ts
@@ -947,6 +947,7 @@ export const aGqlPoolLiquidityBootstrappingV3 = (
             overrides && overrides.hasOwnProperty('isProjectTokenSwapInBlocked')
                 ? overrides.isProjectTokenSwapInBlocked!
                 : false,
+        isSeedless: overrides && overrides.hasOwnProperty('isSeedless') ? overrides.isSeedless! : false,
         lbpName: overrides && overrides.hasOwnProperty('lbpName') ? overrides.lbpName! : 'conscendo',
         lbpOwner: overrides && overrides.hasOwnProperty('lbpOwner') ? overrides.lbpOwner! : 'coma',
         liquidityManagement:
@@ -2104,6 +2105,7 @@ export const aLiquidityBootstrappingPoolV3Params = (
             overrides && overrides.hasOwnProperty('isProjectTokenSwapInBlocked')
                 ? overrides.isProjectTokenSwapInBlocked!
                 : false,
+        isSeedless: overrides && overrides.hasOwnProperty('isSeedless') ? overrides.isSeedless! : false,
         lbpName: overrides && overrides.hasOwnProperty('lbpName') ? overrides.lbpName! : 'consequatur',
         lbpOwner: overrides && overrides.hasOwnProperty('lbpOwner') ? overrides.lbpOwner! : 'culpa',
         projectToken: overrides && overrides.hasOwnProperty('projectToken') ? overrides.projectToken! : 'paulatim',
@@ -2186,6 +2188,8 @@ export const aMutation = (overrides?: Partial<Mutation>): Mutation => {
             overrides && overrides.hasOwnProperty('poolReloadStakingForAllPools')
                 ? overrides.poolReloadStakingForAllPools!
                 : 'anser',
+        poolReloadSwaps:
+            overrides && overrides.hasOwnProperty('poolReloadSwaps') ? overrides.poolReloadSwaps! : 'cubitum',
         poolSyncAllPoolsFromSubgraph:
             overrides && overrides.hasOwnProperty('poolSyncAllPoolsFromSubgraph')
                 ? overrides.poolSyncAllPoolsFromSubgraph!