+++ config/index.ts
@@ -15,6 +15,7 @@ import sonicConfig from './sonic';
 import hyperevmConfig from './hyperevm';
 import plasmaNetworkConfig from './plasma';
 import xlayerNetworkConfig from './xlayer';
+import monadNetworkConfig from './monad';
 
 export const DAYS_OF_HOURLY_PRICES = 100;
 export const BALANCES_SYNC_BLOCKS_MARGIN = 200;
@@ -36,4 +37,5 @@ export default {
     [Chain.HYPEREVM]: hyperevmConfig,
     [Chain.PLASMA]: plasmaNetworkConfig,
     [Chain.XLAYER]: xlayerNetworkConfig,
+    [Chain.MONAD]: monadNetworkConfig,
 };