+++ modules/sources/contracts/v3/fetch-pool-type-data.ts
@@ -6,15 +6,14 @@ import {
     lbpCalls,
     reclammCalls,
 } from '../pool-type-dynamic-data';
-import { multicallViem, ViemMulticallCall } from '../../../web3/multicaller-viem';
 import { PrismaPoolType } from '@prisma/client';
 
-export const poolTypeCalls = (pool: { id: string; type: PrismaPoolType }) => {
+export const poolTypeCalls = (pool: { id: string; type: PrismaPoolType }, vault: string) => {
     switch (pool.type) {
         case PrismaPoolType.STABLE:
             return stableContractCalls(pool.id);
         case PrismaPoolType.LIQUIDITY_BOOTSTRAPPING:
-            return lbpCalls(pool.id);
+            return lbpCalls(pool.id, vault);
         case PrismaPoolType.QUANT_AMM_WEIGHTED:
             return quantAmmWeightedCalls(pool.id);
         case PrismaPoolType.RECLAMM:
@@ -22,19 +21,4 @@ export const poolTypeCalls = (pool: { id: string; type: PrismaPoolType }) => {
         default:
             return [];
     }
-};
-
-export const fetchPoolTypeData = async (
-    client: ViemClient,
-    pools: {
-        id: string;
-        type: PrismaPoolType;
-    }[],
-    blockNumber?: bigint,
-): Promise<{ [address: string]: PoolTypeData }> => {
-    const calls = pools.flatMap(poolTypeCalls).filter((x): x is ViemMulticallCall => !!x);
-
-    const data = await multicallViem(client, calls, blockNumber);
-
-    return data as { [address: string]: PoolTypeData };
 };